refactor: bulk loans
