+++ src/componentsNew/SwapModal/ChangeSidesButton/ChangeSidesButton.tsx
@@ -1,4 +1,5 @@
 import { FC } from 'react';
+
 import styles from './ChangeSidesButton.module.scss';
 
 interface ChangeSidesButtonProps {
