+++ src/pages/BorrowPage/components/SuccessLoan/SuccessLoan.module.scss
@@ -1,0 +1,21 @@
+.wrapper {
+  padding-top: 24px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.title {
+  font: var(--font-text-1);
+  margin-bottom: 12px;
+  text-align: center;
+}
+
+.btn {
+  width: 222px;
+  height: 48px;
+  padding: unset;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
