+++ src/pages/BorrowPage/components/SuccessLoan/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SuccessLoan';
